+++ sdk/type_test.go
@@ -1,4 +1,4 @@
-package types
+package sdk
 
 import (
 	"testing"
